+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 #Changelog
 
 ## latest
+- when setting a slot location: renamed json key from iosLocation to location
+- unify slot media items in one array called 'media' instead of separate arrays
+  for photo, voice and video, added item type info for every item
 
 ### change attributes for Ios Location:
 - remove street, city, postcode, auid