+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 ## latest
 
+
+## v0.5.0 on t2s2-dev
+- when setting a slot location: renamed json key from iosLocation to location
+- unify slot media items in one array called 'media' instead of separate arrays
+  for photo, voice and video, added item type info for every item
+
+### change attributes for Ios Location:
+- remove street, city, postcode, auid
+- add thoroughfare, sub_thoroughfare, locality, sub_locality, administrative_area,
+  sub_administrative_area, postal_code, iso_country_code, in_land_water, ocean,
+  areas_of_interest
+- increase name, country to 255 chars
+
+
 ## v0.4.3 on t2s2-silvio-dev
 - accept local identifier for user image and return it in JSON
 - accept local identifier for group image and return it in JSON